Overhaul settings into a Stripe-style grid of dedicated pages (#344)
Replace the sidebar-menu settings with a landing grid at /settings that
links to one dedicated page per category at /settings/[section]. A single
registry (src/lib/settings/sections.ts) drives both the grid cards and each
page header, grouping every section under three headings (Agent, Workflow &
integrations, System) sized to fit a 1920x1080 screen without scrolling.

Fold in the scoped design decisions:
- Move Automation and Railways off the top nav into settings; the old
  /automation and /railways routes 308-redirect to their /settings homes.
- Group Jira and Tailscale as opt-in integrations under a single Apps page.
- Add a Workspace page grouping agent instructions, the base setup script,
  the config repo, environment variables, network access, and container
  restart/recreate.
- Collect destructive actions into one Danger zone.
- Merge the statistics reset into Usage & stats.
- Remove the settings notepad; the operator notepad already lives on the
  kanban board on the same endpoint.

Reload section data reactively so navigating straight between subpages
refetches, and 404 to a "Not found" header for an unknown section. Update
CLAUDE.md to record the new settings IA.
