+++ frontend/src/routes/+layout.svelte
@@ -21,9 +21,7 @@
     { href: '/suggestions', label: 'Suggestions' },
     { href: '/watch', label: 'Watch' },
     { href: '/compose', label: 'Compose' },
-    { href: '/automation', label: 'Automation' },
     { href: '/repos', label: 'Repositories' },
-    { href: '/railways', label: 'Railways' },
     { href: '/settings', label: 'Settings' }
   ]
